% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br_contrast}
\alias{br_contrast}
\title{Binocular rivalry, variable contrast}
\format{
A data frame with 4616 rows and 6 variables:
\describe{
\item{Observer}{Participant ID.}
\item{Block}{Run / block index.}
\item{Contrast}{Contrast on scale from 0 to 1.}
\item{Time}{Time relative to the run onset in \emph{seconds}}
\item{State}{Factor with levels \code{"Left"}, \code{"Right"} (clear states), and \code{"Mixed"}}.
\item{Duration}{Duration of a dominance phase in \emph{seconds}. Note that the duration for the last dominance phase is curtailed and, therefore, set to zero.}
}
}
\usage{
br_contrast
}
\description{
Dataset on binocular rivalry with variable but equal
contrast for six participants.
}
\keyword{datasets}
