% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br_single_subject}
\alias{br_single_subject}
\title{Single experimental session for binocular rivalry stimulus}
\format{
A data frame with 76 rows and 6 variables:
\describe{
\item{Observer}{Participant ID, all rows contain \emph{"ap"}}
\item{Display}{Display, all rows contain \emph{"BR"}}
\item{Block}{Run / block index}
\item{Time}{Time relative to the run onset in \emph{seconds}}
\item{State}{Index of a perceptually dominant state, \emph{1}, \emph{2} - clear perceptual state, \emph{3} mixed / transition phase}
\item{Duration}{Duration of a dominance phase in \emph{seconds}. Note that the duration for the last dominance phase is curtailed and, therefore, set to zero.}
}
}
\source{
\doi{10.1167/11.10.12}
}
\usage{
br_single_subject
}
\description{
A single subject / multiple runs dataset for binocular rivalry.
}
\keyword{datasets}
