% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br_singleblock}
\alias{br_singleblock}
\title{Single run for binocular rivalry stimulus}
\format{
A data frame with 76 rows and 6 variables:
\describe{
\item{Observer}{Participant ID, all rows contain \emph{"ap"}}
\item{Group}{Display, all rows contain \emph{"BR"}}
\item{Block}{Run / block index, all rows contain \emph{1}}
\item{Time}{Time relative to the run onset in \emph{seconds}}
\item{State}{Index of a perceptually dominant state, \emph{1}, \emph{2} - clear perceptual state, \emph{3} mixed / transition phase}
}
}
\source{
\doi{10.1167/11.10.12}
}
\usage{
br_singleblock
}
\description{
A single subject / single run dataset for binocular rivalry.
}
\keyword{datasets}
