% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fixed_history_parameter.R
\name{check_fixed_history_parameter}
\alias{check_fixed_history_parameter}
\title{Evaluates values for a fixed history parameter}
\usage{
check_fixed_history_parameter(param_name, param_value, randomN, upperLimit)
}
\arguments{
\item{param_name}{Name of the parameter.}

\item{param_value}{A single value or \code{randomN} numeric values.}

\item{randomN}{Number of levels for the random variable.}

\item{upperLimit}{Upper limit for a valid \code{param_value}.}
}
\value{
A numeric vector \code{randomN} long.
}
\description{
Expects either a single value within a valid range or
\code{randomN} values.
}
\examples{
check_fixed_history_parameter("tau", 1, 10, Inf)
}
\keyword{internal}
