% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_parameter.R
\name{check_fixed_or_beta_prior}
\alias{check_fixed_or_beta_prior}
\title{Checks validity of a fixed value or parameters
for a Beta proportion prior distribution. Raises error,
if one of the checks fail.}
\usage{
check_fixed_or_beta_prior(param, param_name)
}
\arguments{
\item{param}{Either a single value or a vector of two parameters for Beta proportion distribution}

\item{param_name}{Name of the parameter for error message}
}
\value{

}
\description{
Checks validity of a fixed value or parameters
for a Beta proportion prior distribution. Raises error,
if one of the checks fail.
}
\examples{
check_fixed_or_beta_prior(0.5, "mixed_state")
check_fixed_or_beta_prior(c(0.5, 1), "mixed_state")
check_fixed_or_beta_prior(-1, "mixed_state")
}
\keyword{internal}
