% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_parameter.R
\name{check_fixed_or_gamma_prior}
\alias{check_fixed_or_gamma_prior}
\title{Checks validity of a fixed value or parameters
for a Gamma prior distribution. Raises error, if one of
the checks fail.}
\usage{
check_fixed_or_gamma_prior(param, param_name)
}
\arguments{
\item{param}{Either a single value or a vector of two parameters for Gamma distribution}

\item{param_name}{Name of the parameter for error message}
}
\value{

}
\description{
Checks validity of a fixed value or parameters
for a Gamma prior distribution. Raises error, if one of
the checks fail.
}
\examples{
check_fixed_or_gamma_prior(1.0, "norm_tau")
check_fixed_or_gamma_prior(c(5, 2), "norm_tau")
check_fixed_or_gamma_prior(-1, "norm_tau")
}
