% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_parameter.R
\name{check_history_init}
\alias{check_history_init}
\title{Checks valid range for initial history values. Returns a vector of two,
thus replicates a scalar to be used for both initial values}
\usage{
check_history_init(param, param_name)
}
\arguments{
\item{param}{Either a single value or a vector of two number in \link{0..1} range}

\item{param_name}{Name of the parameter for error message}
}
\value{
A vector of two, scalar value is repeated.
}
\description{
Checks valid range for initial history values. Returns a vector of two,
thus replicates a scalar to be used for both initial values
}
\keyword{internal}
