% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor_tau.cumhist}
\alias{cor_tau.cumhist}
\title{Computes correlation between observed and
predicted dominance phase duration for each tau.}
\usage{
\method{cor_tau}{cumhist}(cumhist_fit, method = "pearson")
}
\arguments{
\item{cumhist_fit}{A fitted cumhist object.}

\item{method}{String, correlation coefficient to
be computed. Could be \code{"pearson"},
\code{"kendall"}, or \code{"spearman"}. Defaults to
\code{"pearson"}.}
}
\value{
A data.frame with columns \code{normalized_tau}
and \code{R}
}
\description{
Computes correlation between observed and
predicted dominance phase duration for each tau.
}
\examples{
data(br_singleblock)
gamma_fit <- fit_cumhist(br_singleblock, state="State", duration="Duration")
R <- cor_tau(gamma_fit)
plot(R$tau, R$r)

}
