% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumhist-class.R
\docType{class}
\name{cumhist-class}
\alias{cumhist-class}
\alias{cumhist}
\title{Class \code{cumhist}.}
\description{
Cumulative history model fitted to time-series data.
}
\details{
See \code{methods(class = "cumhist")} for an overview of available methods.
}
\section{Slots}{

\describe{
\item{\code{family}}{A \code{string} with distribution family.}

\item{\code{data}}{A \code{list} with preprocessed data.}

\item{\code{stanfit}}{a \code{\link[rstan:stanfit-class]{stanfit}} object.}
}}

\seealso{
\code{\link{fit_cumhist}}
}
