% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_history_parameter.R
\name{evaluate_history_init}
\alias{evaluate_history_init}
\title{Evaluates validity of initial history values.}
\usage{
evaluate_history_init(history_init)
}
\arguments{
\item{history_init}{Either a single value or a pair of values within \link{0..1} range.}
}
\value{
A vector of two values
}
\description{
Checkes number and range of values. If a scalar is passed, uses same value
for both states.
}
\examples{
evaluate_history_init(0.5)
}
\keyword{internal}
