% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_history.R
\name{extract_history}
\alias{extract_history}
\title{Computes history for a fitted model}
\usage{
extract_history(object)
}
\arguments{
\item{object}{An object of class \link[=cumhist-class]{cumhist}}
}
\value{
A matrix of cumulative history values for each state
}
\description{
Computes history for a fitted model, uses only mean values
for each history parameter. Uses values for each random cluster,
if \code{"random"} or \code{"1|random"} parametrisation was used.
}
\examples{
\donttest{
br_fit <- fit_cumhist(br_singleblock, state="State", duration="Duration")
extract_history(br_fit)
}
}
