% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_term.R
\name{extract_replicate_term_to_matrix}
\alias{extract_replicate_term_to_matrix}
\title{Extract a term and replicates it randomN times for each linear model}
\usage{
extract_replicate_term_to_matrix(object, term)
}
\arguments{
\item{object}{An object of class \link[=cumhist-class]{cumhist}}

\item{term}{String, term name}
}
\value{
Matrix
}
\description{
Extract a term and replicates it randomN times for each linear model.
Used for population mean or variance terms.
}
\examples{
\donttest{
br_fit <- fit_cumhist(br_singleblock, state = "State", duration = "Duration")
bH_mu <- extract_replicate_term_to_matrix(br_fit, "bH_mu")
}
}
