% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_term.R
\name{extract_term_to_matrix}
\alias{extract_term_to_matrix}
\title{Extracts a term with one column per fixed or random-level into a matrix}
\usage{
extract_term_to_matrix(object, term)
}
\arguments{
\item{object}{An object of class \link[=cumhist-class]{cumhist}}

\item{term}{String, term name}
}
\value{
Matrix
}
\description{
Extracts a 3D array for a term with  sample, linear-model,
random/fixed-effect order and returns a matrix with samples as rows
and columns in order 1) all random/fixed effects for lm1, 2) all
random/fixed effects for lm2, etc.
}
\examples{
\donttest{
br_fit <- fit_cumhist(br_singleblock, state = "State", duration = "Duration")
a <- extract_term_to_matrix(br_fit, "a")
}
}
