% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumhist.R
\name{fit_cumhist}
\alias{fit_cumhist}
\title{Computes cumulative history for bistable perceptual
rivalry displays.}
\usage{
fit_cumhist(
  data,
  state,
  duration = NULL,
  onset = NULL,
  participant = NULL,
  session = NULL,
  run = NULL,
  mixed_state = 0.5,
  history_mix = 0,
  norm_tau = c(5, 0.2),
  history_init = 0,
  family = "gamma"
)
}
\arguments{
\item{data}{A table with all time-series.}

\item{state}{String, the name of the column that specifies
perceptual state. The column type should be a factor with
two or three levels (the third level is assumed to correspond to a
transition/mixed phase) or should be convertible to a two level
factor (as it would be impossible to infer the identity of transition/
mixed phase).}

\item{duration}{String, name of the column with duration of individual
perceptual dominance phases. Optional, you can specify \code{onset}
instead.}

\item{onset}{String, name of the column with onsets of the perceptual
dominance states. Optional, used to compute duration of the dominance
phases, if these are not provided explicitly via \code{duration}
parameter.}

\item{participant}{String, name of the column that identifies individual participants.
If omitted, the data is assumed to belong to a single participant. If specified and
there is more than one participant, it is used as a random factor
for hierarchical model.}

\item{session}{String, name of the column that identifies unique
experimental session for which a mean dominance phase duration will
be computed (see \code{norm_tau} parameter). Code assumes that session
IDs are different within a participant but can be the same between them.
If omitted, a single mean dominance duration based on the entire time series
is used.}

\item{run}{String, name of the column that identifies unique runs/blocks.
If omitted, the data is assumed to belong to a single time series. Code
assumes that run IDs are different within an experimental session but
can be the same between the session. E.g. session A, runs 1, 2, 3.. and
session B, runs 1, 2, 3 but not session A, runs 1, 2, 1.}

\item{mixed_state}{Specifies an activation level during
transition/mixed phases (state #3, see \code{state}). Either a single
number (range \link{0..1}) that will be used as a fixed level or a vector
of two numbers \code{c(mu, kappa)} that specifies, correspondingly, mean
(range \link{0..1}) and precision (>0) of beta proportion distribution, it
should be sampled from. Defaults to a fixed value of \code{0.5}.}

\item{history_mix}{Specifies how cumulative history levels for
the same and opposite perceptual states are combined when used to
predict dominance phase durations, so that \code{history = history_mix
* history_same + (1 - history_mix) * history_opposite}. Either a single
number (range \link{0..1}) that will be used as a fixed level or a vector
of two numbers \code{c(mu, kappa)} that specifies, correspondingly, mean
(range \link{0..1}) and precision (>0) of beta proportion distribution, it
should be sampled from. Defaults to a fixed value of \code{0}
(predictions are made using cumulative history for the opposite state).}

\item{norm_tau}{Time constant of negative exponential growth/decay
normalized to the mean duration of clear percepts within each \code{session}.
Either a single positive number (>0) that will be used as is or a vector
of two numbers \code{c(shape, scale)} that specifies, correspondingly,
shape and scale of the gamma distribution it should be sampled from.
Defaults to \code{c(5, 0.2)}.}

\item{history_init}{Initial value for cumulative history computation. Either
a numeric scalar in \link{0..1} range or a vector of two numbers in \link{0..1} range.
In the latter case, two histories will start at different levels.}

\item{family}{String, distribution used to fit duration of perceptual dominance
phases. Options include \code{"gamma"} (default), \code{"lognormal"}, \code{"exgauss"}
(exponentially modulated Gaussian), and \code{"normal"}.}
}
\value{

}
\description{
Computes cumulative history for bistable perceptual
rivalry displays.
}
