% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kde_two_observers}
\alias{kde_two_observers}
\title{Multirun data for two participants, kinectic-depth effect display}
\format{
A data frame with 1186 rows and 5 variables:
\describe{
\item{Observer}{Participant ID}
\item{Block}{Run / block index}
\item{State}{Factor variable for state with levels \code{-1} and \code{1} coding two clear perceptual states and \code{-2} the mixed / transition phase}
\item{Time}{Time relative to the run onset in \emph{seconds}}
\item{Duration}{Duration of a dominance phase in \emph{seconds}. Note that the duration for the last dominance phase is curtailed and, therefore, set to zero.}
}
}
\source{
\doi{10.1167/11.10.12}
}
\usage{
kde_two_observers
}
\description{
Multirun data for two participants, kinectic-depth effect display
}
\keyword{datasets}
