% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passing_parameter.R
\name{pass_beta_prop}
\alias{pass_beta_prop}
\title{Packages beta proportion parameter either as fixed or as sampled from priors}
\usage{
pass_beta_prop(param, param_name)
}
\arguments{
\item{param}{A scalar (fixed value) or two number vector (prior for sampling)}

\item{param_name}{String, parameter name}
}
\value{
List with "fit_" + param_name, "fixed_"+param_name,
param_name+"_prior_mu", and param_name+"_prior_kappa"
properties
}
\description{
Packages beta proportion parameter either as fixed or as sampled from priors
}
\examples{
pass_beta_prop(0.5, "mixed_state")
}
\keyword{internal}
