% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passing_parameter.R
\name{pass_gamma}
\alias{pass_gamma}
\title{Packages gamma parameter either as fixed or as sampled from priors}
\usage{
pass_gamma(param, param_name)
}
\arguments{
\item{param}{A scalar (fixed value) or two number vector (prior for sampling)}

\item{param_name}{String, parameter name}
}
\value{
List with "fit_" + param_name, "fixed_"+param_name, and param_name+"_prior"
properties
}
\description{
Packages gamma parameter either as fixed or as sampled from priors
}
\examples{
pass_gamma(1, "norm_tau")
}
\keyword{internal}
