% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cumhist}
\alias{plot.cumhist}
\title{Posterior interval plots for key parameters. Uses bayesplot::mcmc_intervals
routine and supplies better (family-matching) labels for plotted parameters.}
\usage{
\method{plot}{cumhist}(cumhist_fit, ...)
}
\arguments{
\item{cumhist_fit}{A fitted cumhist object.}
}
\value{

}
\description{
Posterior interval plots for key parameters. Uses bayesplot::mcmc_intervals
routine and supplies better (family-matching) labels for plotted parameters.
}
\examples{
data(br_singleblock)
gamma_fit <- bistablehistory::fit_cumhist(br_singleblock, state="State", duration="Duration")
plot(gamma_fit)
}
