% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.cumhist}
\alias{loo.cumhist}
\title{Computes an efficient approximate leave-one-out
cross-validation via loo library. It can be used
for a model comparison via loo::loo_compare() function.}
\usage{
\method{loo}{cumhist}(cumhist_fit)
}
\arguments{
\item{cumhist_fit}{A fitted cumhist object}
}
\value{
A named list, see loo::loo() for details.
}
\description{
Computes an efficient approximate leave-one-out
cross-validation via loo library. It can be used
for a model comparison via loo::loo_compare() function.
}
\examples{
data(br_singleblock)
gamma_fit <- fit_cumhist(br_singleblock, state="State", duration="Duration")
loo_gamma <- loo(gamma_fit)
lognormal_fit <- fit_cumhist(br_singleblock, state="State", duration="Duration", family="lognormal")
loo_lognormal <- loo(lognormal_fit)
loo::loo_compare(loo_gamma, loo_lognormal)
}
