% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cumhist}
\alias{print.cumhist}
\title{Prints out cumhist object}
\usage{
\method{print}{cumhist}(x, ...)
}
\arguments{
\item{x}{A \link[=cumhist-class]{cumhist} object}

\item{...}{Unused}
}
\value{
Nothing, console output only.
}
\description{
Prints out cumhist object
}
\examples{
\donttest{
br_fit <- fit_cumhist(br_singleblock, state="State", duration="Duration", fixed_effects="Time")
br_fit
}
}
