% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic.cumhist}
\alias{waic.cumhist}
\title{Computes widely applicable information criterion
(WAIC) via loo library. It can be used
for a model comparison via loo::loo_compare() function.}
\usage{
\method{waic}{cumhist}(cumhist_fit)
}
\arguments{
\item{cumhist_fit}{A fitted cumhist object}
}
\value{
A named list, see loo::waic() for details.
}
\description{
Computes widely applicable information criterion
(WAIC) via loo library. It can be used
for a model comparison via loo::loo_compare() function.
}
\examples{
data(br_singleblock)
gamma_fit <- fit_cumhist(br_singleblock, state="State", duration="Duration")
waic_gamma <- waic(gamma_fit)
lognormal_fit <- fit_cumhist(br_singleblock, state="State", duration="Duration", family="lognormal")
waic_lognormal <- waic(lognormal_fit)
loo::loo_compare(waic_gamma, waic_lognormal)
}
